+++ internal/entities/dao/dao.go
@@ -34,7 +34,7 @@ type DAO struct {
 	Guidelines       *string           `json:"-"`
 	Template         *string           `json:"-"`
 	ParentID         *string           `json:"parent_id,omitempty"`
-	ActivitySince    *int              `json:"activity_since,omitempty"`
+	ActivitySince    *common.Time      `json:"activity_since,omitempty"`
 	SubscriptionInfo *SubscriptionInfo `json:"subscription_info"`
 }